requirements.txt
# Resume Parser & Hub — Python Dependencies
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
boto3>=1.34.0
pdfplumber>=0.11.0
tqdm>=4.66.0
faiss-cpu>=1.8.0
numpy>=1.24.0
python-dotenv>=1.0.0
